Code cleanup and optimisations
