+++ Models/SARIMAX.py
@@ -4,50 +4,37 @@ from statsmodels.tsa.statespace.sarimax import SARIMAX
 import matplotlib.pyplot as plt
 from sklearn.metrics import mean_squared_error
 
-df = pd.read_csv('../Data/ICAC Data.csv', parse_dates=['Date'], dayfirst=True)
-df.set_index('Date', inplace=True)
-df.sort_index(inplace=True)
 
-lags = [1, 2, 3, 7, 14, 21, 28]
-for lag in lags:
-    df[f"lag_{lag}"] = df['Output'].shift(lag)
+def execute_sarimax(data):
+    train_size = int(0.8 * len(data))
+    train, test = data.iloc[:train_size], data.iloc[train_size:]
 
-window_sizes = [7, 14, 21, 28]
-for window in window_sizes:
-    df[f"rolling_mean_{window}"] = df['Output'].rolling(window=window).mean()
-    df[f"rolling_std_{window}"] = df['Output'].rolling(window=window).std()
+    X_train, X_test = train.drop("Output", axis=1), test.drop("Output", axis=1)
+    y_train, y_test = train["Output"], test["Output"]
 
-df_fe = df.dropna()
+    sarimax_model = SARIMAX(y_train, exog=X_train, order=(1, 1, 1), seasonal_order=(1, 1, 1, 365))
+    sarimax_result = sarimax_model.fit(disp=False)
 
-train_size = int(0.8 * len(df_fe))
-train, test = df_fe.iloc[:train_size], df_fe.iloc[train_size:]
+    sarimax_predictions = sarimax_result.predict(start=y_test.index[0], end=y_test.index[-1], exog=X_test)
 
-X_train, X_test = train.drop("Output", axis=1), test.drop("Output", axis=1)
-y_train, y_test = train["Output"], test["Output"]
+    sarimax_rmse = np.sqrt(mean_squared_error(y_test, sarimax_predictions))
+    print(f"RMSE for SARIMAX model: {sarimax_rmse:.4f}")
 
-sarimax_model = SARIMAX(y_train, exog=X_train, order=(1, 1, 1), seasonal_order=(1, 1, 1, 365))
-sarimax_result = sarimax_model.fit(disp=False)
+    forecast_data = data.iloc[-len(y_test):].copy()
+    exog_forecast = forecast_data[X_train.columns]
+    sarimax_forecast_45 = sarimax_result.get_forecast(steps=45, exog=exog_forecast).predicted_mean
 
-sarimax_predictions = sarimax_result.predict(start=y_test.index[0], end=y_test.index[-1], exog=X_test)
+    print(sarimax_forecast_45)
 
-sarimax_rmse = np.sqrt(mean_squared_error(y_test, sarimax_predictions))
-print(f"RMSE for SARIMAX model: {sarimax_rmse:.4f}")
-
-forecast_df = df_fe.iloc[-len(y_test):].copy()
-exog_forecast = forecast_df[X_train.columns]
-sarimax_forecast_45 = sarimax_result.get_forecast(steps=45, exog=exog_forecast).predicted_mean
-
-print(sarimax_forecast_45)
-
-plt.figure(figsize=(16, 8))
-plt.plot(df['Output'], label="Actual Data", color="blue")
-plt.plot(y_test.index, sarimax_predictions, label="SARIMAX Predictions", color="red", linestyle="--")
-forecast_index_45 = pd.date_range(df.index[-1], periods=46, inclusive='right')
-plt.plot(forecast_index_45, sarimax_forecast_45, 'g--', label="45-day SARIMAX Forecast", alpha=0.7)
-plt.title("Cotlook A Index with SARIMAX Predictions and 45 Days Forecast")
-plt.xlabel("Date")
-plt.ylabel("Cotlook A Index")
-plt.legend()
-plt.grid(True)
-plt.tight_layout()
-plt.show()
+    plt.figure(figsize=(16, 8))
+    plt.plot(data['Output'], label="Actual Data", color="blue")
+    plt.plot(y_test.index, sarimax_predictions, label="SARIMAX Predictions", color="red", linestyle="--")
+    forecast_index_45 = pd.date_range(data.index[-1], periods=46, inclusive='right')
+    plt.plot(forecast_index_45, sarimax_forecast_45, 'g--', label="45-day SARIMAX Forecast", alpha=0.7)
+    plt.title("Cotlook A Index with SARIMAX Predictions and 45 Days Forecast")
+    plt.xlabel("Date")
+    plt.ylabel("Cotlook A Index")
+    plt.legend()
+    plt.grid(True)
+    plt.tight_layout()
+    plt.show()
